Export as.igraph as a S3 method on Node
`as.igraph` was not automatically registered in R-devel (to be R 4.0.0)
when running Travis-CI checks (but was OK in R 3.5.1 and 3.6.1).

Ideally, we would only export `as.igraph` if the user has the package
`igraph` installed.

This commit provides a simple fix: we Import 'igraph' and export it's
generic `as.igraph`; then we add the S3 method `as.igraph.Node`. The
`node` argument to `as.igraph.Node` had to be renamed to match
`igraph::as.igraph`s formals (`node` -> `x`; include a `...`
argument).